YTI-879 - getting rid of the term classification.
This commit contains backend refactorings where the terms classication
and dataClassification(s) are getting removed. So far, only various
Java objects have been touched, that is the names of variables, methods,
and classes. No constants or paths or string literals (of the ES index
for example) have been touched.  Also, no changes to the UI yet.