update docker-maven-plugin for integration tests
Previous commit removed spotify's docker-maven-plugin for building,
but fabric8's docker-maven-plugin still remains for integration tests.

This commit bumps up the plugin version to 0.38.1, which should work
with docker using a socket file.

Elasticsearch version is also updated to the same (but still old)
version that is used by the docker-compose setup in the project.